+++ Source/Examples/DemoLauncher/ExampleInfo.cs
@@ -8,7 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows;
-using PropertyTools;
+using ExampleLibrary;
 
 namespace DemoLauncher
 {