+++ Source/Examples/PropertyGrid/PropertyGridDemos/Examples/GettingStartedExample.xaml.cs
@@ -10,6 +10,7 @@
 namespace PropertyGridDemos
 {
     using PropertyTools;
+    using ExampleLibrary;
 
     using System.ComponentModel;